backend/requirements.txt
fastapi
uvicorn[standard]
httpx
numpy
pandas
python-dotenv
apscheduler
